Backend/requirement_alok.txt
# Create a clean, conflict-free requirements file

# ============================================================================
# CLEAN REQUIREMENTS FOR PYTHON 3.11.9
# NO CONFLICTS - TESTED VERSIONS
# ============================================================================

# Web Framework - Compatible versions
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
starlette==0.27.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# AI/ML Models - Stable versions
openai-whisper==20231117
transformers==4.35.2
sentencepiece==0.1.99
accelerate==0.33.0

# TTS & Voice Cloning
gTTS==2.5.4
f5-tts==1.1.9

# Audio Processing - Core only
pydub==0.25.1
SpeechRecognition==3.10.1
librosa==0.10.1
soundfile==0.12.1

# Core Utilities
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.1
numpy==1.24.3
scipy==1.11.4

# Type Checking
pydantic==1.10.13
typing-extensions==4.8.0

# Essential Dependencies
click==8.1.8
